// src/app/type/form-field-types.enum.ts
export enum FormFieldTypes {
  Tme = 'TME',
  Date = 'Date',
  Text = 'Text',
  Number = 'Number',
  RichText = 'Rich Text',
  PassFail = 'Passed/Failed',
  StaticText = 'Static Text',
  StaticRichText = 'Static Rich Text',
  MultipleOptions = 'Multiple Options',
  DigitalSignature = 'Digital Signature',
  RadioButtonGroup = 'Radio Button Group',
}